refactor: Extract javascript bridges to their own file
